requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.5
google-cloud-storage==2.10.0
yt-dlp==2025.7.21
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6